adding position and normal resolution

+++ include/descriptorset.h
@@ -40,7 +40,7 @@ struct light_ubo_t
 struct render_ubo_t
 {
     glm::vec4 ambient = glm::vec4(0.2f);
-    int sampleCount = 2;
+    float sampleCount = 2.0f;
 };
 
 struct post_ubo_t
